pyproject.toml
[project]
name = "pybantic"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [{ name = "dongruixiao", email = "dongruixiao@gmail.com" }]
requires-python = ">=3.11"
dependencies = [
    "grpcio>=1.71.0",
    "grpcio-tools>=1.71.0",
    "jinja2>=3.1.6",
    "protobuf>=5.29.4",
    "pydantic>=2.11.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = ["ipython>=9.2.0"]

[tool.uv.workspace]
members = ["samples/hello_world"]
